+++ src/provider/__tests__/ClashProvider.test.ts
@@ -143,8 +143,6 @@ test('getClashSubscription', async (t) => {
     type: NodeTypeEnum.Vmess,
     nodeName: 'vmess new format',
     hostname: 'server',
-    host: 'v2ray.com',
-    path: '/path',
     port: 443,
     uuid: 'uuid',
     alterId: '32',
@@ -154,16 +152,19 @@ test('getClashSubscription', async (t) => {
     tls: true,
     tls13: false,
     skipCertVerify: true,
-    wsHeaders: {
-      host: 'v2ray.com',
+    wsOpts: {
+      headers: {
+        Host: 'v2ray.com',
+      },
+      path: '/path',
+      'early-data-header-name': 'Sec-WebSocket-Protocol',
+      'max-early-data': 2048,
     },
   })
   t.deepEqual(config.shift(), {
     type: NodeTypeEnum.Vmess,
     nodeName: 'vmess custom header',
     hostname: 'server',
-    host: 'server',
-    path: '/path',
     port: 443,
     uuid: 'uuid',
     alterId: '32',
@@ -173,8 +174,11 @@ test('getClashSubscription', async (t) => {
     tls: true,
     tls13: false,
     skipCertVerify: false,
-    wsHeaders: {
-      edge: 'www.baidu.com',
+    wsOpts: {
+      headers: {
+        edge: 'www.baidu.com',
+      },
+      path: '/path',
     },
   })
   t.deepEqual(config.shift(), {
@@ -344,6 +348,121 @@ foo: bar
   )
 
   scope.done()
+})
+
+test('vmess Configurations', (t) => {
+  t.deepEqual(
+    parseClashConfig([
+      {
+        type: 'vmess',
+        name: 'vmess',
+        server: 'server',
+        port: 443,
+        uuid: 'uuid',
+        alterId: 32,
+        cipher: 'auto',
+        network: 'http',
+        'http-opts': {
+          path: ['/path'],
+          headers: {
+            host: ['v2ray.com'],
+          },
+        },
+      },
+    ]),
+    [
+      {
+        type: NodeTypeEnum.Vmess,
+        nodeName: 'vmess',
+        hostname: 'server',
+        port: 443,
+        uuid: 'uuid',
+        alterId: '32',
+        method: 'auto',
+        network: 'http',
+        tls: false,
+        udpRelay: false,
+        httpOpts: {
+          path: ['/path'],
+          headers: {
+            host: 'v2ray.com',
+          },
+        },
+      },
+    ],
+  )
+
+  t.deepEqual(
+    parseClashConfig([
+      {
+        type: 'vmess',
+        name: 'vmess',
+        server: 'server',
+        port: 443,
+        uuid: 'uuid',
+        alterId: 32,
+        cipher: 'auto',
+        network: 'grpc',
+        'grpc-opts': {
+          'grpc-service-name': 'service',
+        },
+      },
+    ]),
+    [
+      {
+        type: NodeTypeEnum.Vmess,
+        nodeName: 'vmess',
+        hostname: 'server',
+        port: 443,
+        uuid: 'uuid',
+        alterId: '32',
+        method: 'auto',
+        network: 'grpc',
+        tls: false,
+        udpRelay: false,
+        grpcOpts: {
+          serviceName: 'service',
+        },
+      },
+    ],
+  )
+
+  t.deepEqual(
+    parseClashConfig([
+      {
+        type: 'vmess',
+        name: 'vmess',
+        server: 'server',
+        port: 443,
+        uuid: 'uuid',
+        alterId: 32,
+        cipher: 'auto',
+        network: 'h2',
+        'h2-opts': {
+          path: '/path',
+          host: ['v2ray.com'],
+        },
+      },
+    ]),
+    [
+      {
+        type: NodeTypeEnum.Vmess,
+        nodeName: 'vmess',
+        hostname: 'server',
+        port: 443,
+        uuid: 'uuid',
+        alterId: '32',
+        method: 'auto',
+        network: 'h2',
+        tls: false,
+        udpRelay: false,
+        h2Opts: {
+          path: '/path',
+          host: ['v2ray.com'],
+        },
+      },
+    ],
+  )
 })
 
 test('snell Configurations', (t) => {